Make more functions part of the base API (#156)
`n_1d_arrays`, `check_fit_input` and `least_squares` are now part of the
`verde.base` API. These functions are needed for building new gridders.
Reorganize `verde.base` into a package and improve the documentation for
`least_squares`.

Fixes #154
