+++ verde/base/__init__.py
@@ -1,0 +1,4 @@
+# pylint: disable=missing-docstring
+from .base_classes import BaseGridder
+from .least_squares import least_squares
+from .utils import n_1d_arrays, check_fit_input
