+++ verde/spline.py
@@ -6,9 +6,9 @@ from warnings import warn
 import numpy as np
 from sklearn.utils.validation import check_is_fitted
 
-from .base import BaseGridder, check_fit_input, least_squares
+from .base import n_1d_arrays, BaseGridder, check_fit_input, least_squares
 from .coordinates import get_region
-from .utils import n_1d_arrays, parse_engine
+from .utils import parse_engine
 
 try:
     import numba
