+++ verde/trend.py
@@ -4,9 +4,8 @@
 import numpy as np
 from sklearn.utils.validation import check_is_fitted
 
-from .base import BaseGridder, check_fit_input, least_squares
+from .base import BaseGridder, check_fit_input, least_squares, n_1d_arrays
 from .coordinates import get_region
-from .utils import n_1d_arrays
 
 
 class Trend(BaseGridder):
